Major refactor of node pooling strategy.
Node connectors and connections are now keyspace aware. You no longer
have to specify an explicit keyspace on connecting to a cluster, but
commands to the cluster are now bound to a particular keyspace. Node connectors
are smart about allocating connections to a given keyspace.

Additionally, fixed a couple bugs in the Session interface and also
added a bit of initial sanity checking around the existence of a keyspace
when constructing a session.

Also bumped the version to 0.2.0-SNAPSHOT due to a couple API changes.